requirements.txt
# Core packages
numpy==1.24.3
pandas==2.0.2
scipy==1.10.1

# Data acquisition
alpha_vantage==2.3.1
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.10.0
yfinance==0.2.12

# Machine learning and data science
tensorflow==2.13.0
scikit-learn==1.3.0
keras==2.13.1
xgboost==1.7.6
lightgbm==3.3.5

# NLP and text processing
nltk==3.8.1
textblob==0.17.1
transformers==4.30.2

# Database
sqlalchemy==2.0.19
pymysql==1.1.0

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
dash>=2.11.1

# Web
Flask>=2.0.0,<2.3.0
streamlit==1.24.1

# Testing and quality
pytest==7.4.0
pylint==2.17.5

# Utilities
tqdm==4.65.0
pyyaml==6.0
python-dotenv==1.0.0
joblib==1.3.1 